+++ R/clean_names.R
@@ -70,17 +70,21 @@ clean_names <- function(dat, ...) {
 
 #' @rdname clean_names
 #' @export
-clean_names.data.frame <- function(dat, ...) {
-  stats::setNames(dat, make_clean_names(names(dat), ...))
-}
-
-#' @rdname clean_names
-#' @export
 clean_names.default <- function(dat, ...) {
-  stop(
-    "No `clean_names()` method exists for the class ", paste(class(dat), collapse=", "),
-    "\nConsider janitor::make_clean_names() for other cases of manipulating vectors of names."
-  )
+  if(is.null(names(dat)) && is.null(dimnames(dat))) {
+    stop(
+      "`clean_names()` requires that either names or dimnames be non-null.",
+      call. = FALSE
+    )
+    return(dat)
+  }
+  if(is.null(names(dat))) {
+    colnames(dat) <- make_clean_names(colnames(dat), ...)
+    rownames(dat) <- make_clean_names(rownames(dat), ...)
+  } else {
+    names(dat) <- make_clean_names(names(dat), ...)
+  }
+  dat
 }
 
 #' @rdname clean_names